cleanup: removed misc code and seperated google components into own seperate file
