+++ src/routes.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import Homepage from './components/Homepage/Homepage';
 import SignUp from './components/SignUp/SignUp';
 import Login from './components/Login/Login';
-import testGoogleMap from './components/PlacesAutocomplete/GoogleMapsTest'
+
 
 import Profile from './components/Profile/Profile';
 import PropertySearch from './components/PropertySearch/PropertySearch';
@@ -21,7 +21,7 @@ class Routes extends Component {
         <Route exact path="/login" component={Login} />
         <Route exact path="/profile" component={Profile} />
         <Route exact path="/propertysearch" component={PropertySearch} />
-        <Route exact path="/test" component={testGoogleMap} />
+
 
         {/* <Route exact path="/contact" component={Contact} /> */}
       </Switch>
